% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_blocks_functions.R
\name{average_partition_matrix}
\alias{average_partition_matrix}
\title{average_partition_matrix}
\usage{
average_partition_matrix(repeat_result)
}
\arguments{
\item{repeat_result}{A list containing the results from `repeat_sample_likely_partition`.}
}
\value{
A matrix representing the average partition likelihood between nodes.
}
\description{
Calculate the average partition matrix from multiple runs of the
`repeat_sample_likely_partition` function.
This function takes the results from `repeat_sample_likely_partition`
and computes a matrix indicating the average likelihood of each pair of nodes
being in the same partition.
}
\examples{
# Create a square matrix with 40 rows and 40 columns
test_mat <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Repeat sampling a likely partition of the matrix with 2 blocks for 10 runs
repeat_result <- repeat_sample_likely_partition(test_mat, n_blocks = 2, n_runs = 10, n_iter = 1000)
avg_matrix <- average_partition_matrix(repeat_result)
}
