% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{calculate_log_linear_model}
\alias{calculate_log_linear_model}
\title{calculate_log_linear_model}
\usage{
calculate_log_linear_model(mobility_table, partition, include_coeffs = FALSE)
}
\arguments{
\item{mobility_table}{A square matrix representing the mobility table.}

\item{partition}{A vector representing the partition of the square matrix
into blocks.}

\item{include_coeffs}{A boolean indicating whether to return the
coefficients of the model.}
}
\value{
A list containing the coefficients of the model (if `include_coeffs` is TRUE)
and the fit of the model (AIC) - if `include_coeffs` is FALSE,
the returned fit is includes an unknown constant.
}
\description{
Calculate the parameters of a log-linear model
in which block membership predicts cell counts of the
square matrix;
optionally return the coefficients of the model
}
\examples{
# Create a square matrix with 40 rows and 40 columns
test_mat <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Partition the matrix into 2 blocks
partition <- partition_square_matrix(test_mat, n_blocks = 2)
# Calculate the log-linear model
calculate_log_linear_model(test_mat, partition, include_coeffs = TRUE)
}
