% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{partition_square_matrix}
\alias{partition_square_matrix}
\title{partition_square_matrix}
\usage{
partition_square_matrix(mobility_table, n_blocks)
}
\arguments{
\item{mobility_table}{A square matrix representing the mobility table.}

\item{n_blocks}{The number of blocks to partition the matrix into.}
}
\value{
A vector representing the partition of the square matrix into blocks.
}
\description{
Randomly partition a square matrix into a specified number of
(approximately) equal sized blocks.
}
\examples{
# Create a square matrix with 40 rows and 40 columns
test_mat <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Partition the matrix into 2 blocks
partition_square_matrix(test_mat, n_blocks = 2)
}
