% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_change_functions.R
\name{propose_multipartite_partition_update_swap}
\alias{propose_multipartite_partition_update_swap}
\title{propose_multipartite_partition_update_swap}
\usage{
propose_multipartite_partition_update_swap(
  table_aa,
  table_ab,
  table_bb,
  partition,
  swap_a = NULL
)
}
\arguments{
\item{table_aa}{A square matrix representing the mobility table
in which rows and columns are partitioned into the same blocks (a).}

\item{table_ab}{A square matrix representing the mobility table
in which rows are partitioned into blocks (a) and columns are partitioned
into different blocks (b).}

\item{table_bb}{A square matrix representing the mobility table
in which rows and columns are partitioned into the same blocks (b).}

\item{partition}{A list containing two vectors:
one for the partitioning of (a)
and one for the partitioning of (b).}

\item{swap_a}{A boolean indicating whether to swap the block membership of blocks (a) (TRUE)
or blocks (b) (FALSE).}
}
\value{
A list containing the updated partitions (a) and (b)
and the summed fit of the log-linear model for the three matrices.
}
\description{
Propose an update of one out of two given partitions
by swapping the block membership of two nodes.
The partitions guide three matrices:
1. A sqare matrix in which rows and columns are partitioned
into the same blocks (a),
2. A square matrix in which rows are partitioned into blocks (a)
and columns are partitioned into different blocks (b),
3. A square matrix in which rows and columns are partitioned
into the same blocks (b).
Accept the update with a probability that is proportional to the
exp() of the difference in summed fit of the log-linear model
of the three matrices comparing the current partitioning
and the proposed partition.
In each updating step, this function will either swap the
block membership of blocks (a) or blocks (b),
depending on a parameter or randomly.
}
\examples{
# Create a square matrix with 40 rows and 40 columns
mat_aa <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Create a square matrix with 40 rows and 40 columns
mat_ab <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Create a square matrix with 40 rows and 40 columns
mat_bb <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Partition the matrices into 2 blocks for (a) and 3 blocks for (b)
init_partition <- partition_square_matrix(mat_ab, n_blocks = c(2, 3))
# Propose an update of the partition by swapping nodes
propose_multipartite_partition_update_swap(mat_aa, mat_ab, mat_bb,
 init_partition, swap_a = TRUE)
}
