% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_change_functions.R
\name{repeat_sample_likely_partition}
\alias{repeat_sample_likely_partition}
\title{repeat_sample_likely_partition}
\usage{
repeat_sample_likely_partition(
  mobility_table,
  n_blocks = 2,
  n_runs = 10,
  n_iter = 100
)
}
\arguments{
\item{mobility_table}{A square matrix representing the mobility table.}

\item{n_blocks}{The number of blocks to partition the matrix into.}

\item{n_runs}{The number of runs to perform.}

\item{n_iter}{The number of iterations to run the sampling in each run.}
}
\value{
A list containing two lists (partitions and fits), where each list contains the
partition and fit from each run; as well as the mobility table.
}
\description{
Repeat the function sample_likely_partition for a specified number of runs
the first time with a random partition
and then with the partition from the previous run
return the partition and the fit of each run as two lists
}
\examples{
# Create a square matrix with 40 rows and 40 columns
test_mat <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Repeat sampling a likely partition of the matrix with 2 blocks for 10 runs
repeat_sample_likely_partition(test_mat, n_blocks = 2, n_runs = 10, n_iter = 1000)
}
