% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_blocks_functions.R
\name{repeat_until_no_improvement}
\alias{repeat_until_no_improvement}
\title{repeat_until_no_improvement}
\usage{
repeat_until_no_improvement(
  mobility_table,
  init_partition = NULL,
  n_blocks = 2,
  max_iter = 100
)
}
\arguments{
\item{mobility_table}{A square matrix representing the mobility table.}

\item{init_partition}{An optional initial partition vector.}

\item{n_blocks}{The number of blocks to partition the matrix into.}

\item{max_iter}{The maximum number of iterations to perform.}
}
\value{
A list containing the final partition and the fit of the model.
}
\description{
Repeatedly update the partition of a square matrix
until no further improvement in fit is possible.
This function uses the `update_partition_node_swap` and
`update_partition_node_move` functions to
improve the partition iteratively.
}
\examples{
# Create a square matrix with 40 rows and 40 columns
test_mat <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Update partition and repeat until no improvement in fit is possible
repeat_until_no_improvement(test_mat, n_blocks = 2, max_iter = 100)
}
