% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_change_functions.R
\name{sample_likely_partition}
\alias{sample_likely_partition}
\title{sample_likely_partition}
\usage{
sample_likely_partition(
  mobility_table,
  initial_partition = NULL,
  n_blocks = 2,
  n_iter = 1000
)
}
\arguments{
\item{mobility_table}{A square matrix representing the mobility table.}

\item{initial_partition}{A vector representing the initial partition of the
square matrix into blocks. If NULL, a random partition will be created.}

\item{n_blocks}{The number of blocks to partition the matrix into.
If two numbers are provided, the first number is the number of blocks for rows
and the second number is the number of blocks for columns.}

\item{n_iter}{The number of iterations to run the sampling.}
}
\value{
A list containing the sampled partition and the fit of the log-linear model.
}
\description{
Sample a likely partition of the square matrix
Use either propose_partition_update_move or propose_partition_update_swap,
randomly for a specified number of iterations
and return the partition and the fit
}
\examples{
# Create a square matrix with 40 rows and 40 columns
test_mat <- matrix(rpois(1600, lambda = 10), nrow = 40, ncol = 40)
# Sample a likely partition of the matrix with 2 blocks
sample_likely_partition(test_mat, n_blocks = 2, n_iter = 1000)
}
