% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocBM-package.R
\docType{package}
\name{stocBM-package}
\alias{stocBM}
\alias{stocBM-package}
\title{stocBM: Stochastic Blockmodelling for Mobility Networks}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Per Block \email{block@soziologie.uzh.ch} (\href{https://orcid.org/0000-0002-7583-2392}{ORCID}) [copyright holder]

}
\keyword{internal}
